services/dialogue_worker/requirements.txt
pypdf2>=3.0.0
python-docx>=1.0.0
minio>=7.1.0
aiohttp>=3.8.0
nats-py>=2.3.0 